refactor: centralize duplicated API request/response types [#127]
The classification and official-form-lookup API contracts were redefined
locally in hooks, components, and the route handler, with inconsistent
strictness (confidence optional/missing, issueType string vs enum). A
contract change had to be hand-propagated to several files.

- Add src/lib/api/types.ts as the shared home for over-the-wire API
  contract types; define the official-form-lookup result there once.
- Consolidate the route's FormLinkResult and the hook/component's
  OfficialFormLookupResult (structurally identical) into the shared type;
  the route, use-form-lookup, and review-step all import it.
- Point the classification types at their existing source of truth in
  src/lib/classify/types.ts: use-report-submission and review-step now
  import the canonical ClassificationResult / ProviderResult /
  ComparisonResult (the shape the classify route already returns) instead
  of looser local copies.

No runtime behavior or response shape changes.
